requirements.txt
mypy==1.14.0
sqlalchemy
uvicorn==0.34.0
requests
python-dotenv==1.0.1
pydantic==2.10.4
python-dotenv
google-ads==22.0.0
pydantic
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
google-cloud-storage==2.19.0
python-multipart==0.0.20
fastapi==0.115.6
fastapi